Move section to own component
