+++ ui/src/Pages/Preferences/Appearance/Sidebar.jsx
@@ -1,0 +1,14 @@
+import Toggle from "../../../Components/Toggle";
+
+function Sidebar() {
+  return (
+    <section>
+      <h2>Sidebar</h2>
+      <Toggle
+        name="Keep the sidebar always in compact mode"
+      />
+    </section>
+  );
+}
+
+export default Sidebar;
