+++ ui/src/Pages/Preferences/Player/Audio.jsx
@@ -1,0 +1,8 @@
+const Audio = () => (
+  <section>
+    <h2>Audio</h2>
+    {/* <p className="desc">Manage your default preferred subtitle language</p> */}
+  </section>
+);
+
+export default Audio;
