+++ ui/src/Pages/Preferences/Player/Index.jsx
@@ -1,29 +1,13 @@
-import { useCallback } from "react";
-import Toggle from "../../../Components/Toggle";
+import Subtitles from "./Subtitles";
+import Audio from "./Audio";
 
 import "./Index.scss";
 
-function VideoPlayer() {
-  const toggleShowSubsDefault = useCallback((state) => {
-    console.log(state);
-  }, []);
+const Player = () => (
+  <div className="preferencesPlayer">
+    <Subtitles/>
+    <Audio/>
+  </div>
+);
 
-  return (
-    <div className="preferencesVideoPlayer">
-      <section>
-        <h2>Subtitles</h2>
-        <Toggle
-          disabled
-          name="Show subtitles by default"
-          onToggle={toggleShowSubsDefault}
-        />
-      </section>
-      <section>
-        <h2>Audio</h2>
-        {/* <p className="desc">Manage your default preferred subtitle language</p> */}
-      </section>
-    </div>
-  );
-}
-
-export default VideoPlayer;
+export default Player;
