+++ ui/src/Pages/Preferences/Player/Index.scss
@@ -1,4 +1,4 @@
-.preferencesVideoPlayer {
+.preferencesPlayer {
   display: grid;
   gap: 1em;
 }
